Move XML test utilities to corehq.tests.util.xml
